Make localStoragePlugin behave correctly in an SSR context
localStorage isn't available on the server in a SSR Nuxt app, but we still want the client to come in and restore the form state.

This change makes the server skip loading the plugin (and not crash on the server due to lack of localStorage access), allowing the client side to handle this properly.

Co-authored-by: Andrew Boyd <andrew@formkit.com>